refactoring: decompose theme switch logic
set new useSwitchTheme hook to provide theme state, move ITheme to models, move themes array to constants
