+++ client/src/components/ThemeSwitcher.tsx
@@ -1,61 +1,19 @@
-import React, {memo, useEffect, useState} from 'react';
+import React, {memo, useEffect} from 'react';
 import {Listbox} from "@headlessui/react";
-import {ComputerDesktopIcon, MoonIcon, SunIcon} from "@heroicons/react/24/outline";
-import {IIcon} from "../types/IIcon";
+import {MoonIcon, SunIcon} from "@heroicons/react/24/outline";
 import {ChevronDownIcon} from "@heroicons/react/20/solid";
+import {useSwitchTheme} from "../hooks/useThemeSwitch";
+import {themes} from "../constants/themes";
 
-interface ITheme {
-    id: number,
-    mode: "light" | "dark" | "system"
-    Icon: React.FC<IIcon>
-}
-
-const themes: Array<ITheme> = [
-    {id: 1, mode: 'light', Icon: SunIcon},
-    {id: 2, mode: 'dark', Icon: MoonIcon},
-    {id: 3, mode: 'system', Icon: ComputerDesktopIcon},
-]
-
-const TestThemeSwitcher = memo(() => {
-    const [selectedTheme, setSelectedTheme] = useState(themes[0])
-    const switchTheme = (theme: ITheme) => {
-        switch (theme.mode) {
-            case "light":
-                localStorage.theme = "light"
-                document.documentElement.classList.add('light')
-                document.documentElement.classList.remove('dark')
-                document.documentElement.style.backgroundColor = "#ffffff"
-                setSelectedTheme(themes[0])
-                break
-            case "dark":
-                localStorage.theme = 'dark'
-                document.documentElement.classList.add('dark')
-                document.documentElement.classList.remove('light')
-                document.documentElement.style.backgroundColor = "#0f172a"
-                setSelectedTheme(themes[1])
-                break
-            default:
-                localStorage.theme = "system"
-                if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
-                    document.documentElement.classList.add('dark')
-                    document.documentElement.classList.remove('light')
-                    document.documentElement.style.backgroundColor = "#0f172a"
-                    setSelectedTheme(themes[2])
-                } else {
-                    document.documentElement.classList.add('light')
-                    document.documentElement.classList.remove('dark')
-                    document.documentElement.style.backgroundColor = "#f3f4f6"
-                    setSelectedTheme(themes[2])
-                }
-        }
-    }
+const ThemeSwitcher = memo(() => {
+    const {selectedTheme, setSelectedTheme} = useSwitchTheme()
     useEffect(() => {
-        switchTheme(themes.find(theme => theme.mode === localStorage.theme) || themes[2])
+        setSelectedTheme(themes.find(theme => theme.mode === localStorage.theme) || themes[2])
     }, [])
 
     return (
         <div className={"relative"}>
-            <Listbox value={selectedTheme} onChange={switchTheme}>
+            <Listbox value={selectedTheme} onChange={setSelectedTheme}>
                 <Listbox.Button
                     className={"flex items-center border border-gray-300 h-10 px-4 space-x-1.5 text-slate-600 rounded-3xl hover:bg-gray-200/30 dark:hover:bg-slate-500/50 dark:border-slate-600 dark:text-slate-300 dark:bg-slate-800"}>
                     <span className={"capitalize text-sm"}>{localStorage.theme}</span>
@@ -95,4 +53,4 @@ const TestThemeSwitcher = memo(() => {
     );
 });
 
-export default TestThemeSwitcher;
+export default ThemeSwitcher;
